Added test and renamed new chimera file
